refactor: move FungibleFaucetError to miden-lib
