+++ crates/miden-objects/src/errors.rs
@@ -97,8 +97,6 @@ pub enum AccountError {
     BuildError(String, #[source] Option<Box<AccountError>>),
     #[error("failed to parse account ID from final account header")]
     FinalAccountHeaderIdParsingFailed(#[source] AccountIdError),
-    #[error("failed to create basic fungible faucet")]
-    FungibleFaucetError(#[source] FungibleFaucetError),
     #[error("account header data has length {actual} but it must be of length {expected}")]
     HeaderDataIncorrectLength { actual: usize, expected: usize },
     #[error("new account nonce {new} is less than the current nonce {current}")]
@@ -134,22 +132,6 @@ pub enum AccountError {
     /// this error type.
     #[error("assumption violated: {0}")]
     AssumptionViolated(String),
-}
-
-#[derive(Debug, Error)]
-pub enum FungibleFaucetError {
-    #[error("faucet metadata decimals is {actual} which exceeds max value of {max}")]
-    TooManyDecimals { actual: u64, max: u8 },
-    #[error("faucet metadata max supply is {actual} which exceeds max value of {max}")]
-    MaxSupplyTooLarge { actual: u64, max: u64 },
-    #[error(
-        "account interface provided for faucet creation does not have basic fungible faucet component"
-    )]
-    NoAvailableInterface,
-    #[error("storage offset `{0}` is invalid")]
-    InvalidStorageOffset(u8),
-    #[error("invalid token symbol")]
-    InvalidTokenSymbol(#[source] TokenSymbolError),
 }
 
 // ACCOUNT ID ERROR
